Compliance fun for @frasergreenroyd
