+++ Physical_Engine/Physical_Engine.csproj
@@ -67,8 +67,8 @@
     <Compile Include="Create\Bracing.cs" />
     <Compile Include="Create\Cable.cs" />
     <Compile Include="Create\Column.cs" />
-    <Compile Include="Create\NewElement1D.cs" />
-    <Compile Include="Create\NewInternalElement2D.cs" />
+    <Compile Include="Create\IElement\IElement1D\NewElement1D.cs" />
+    <Compile Include="Create\IElement\IElement2D\NewInternalElement2D.cs" />
     <Compile Include="Create\Pile.cs" />
     <Compile Include="Create\Roof.cs" />
     <Compile Include="Create\Beam.cs" />
@@ -85,6 +85,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\ConstructionByName.cs" />
     <Compile Include="Query\ExternalPolyline.cs" />
+    <Compile Include="Query\InternalElements2D.cs" />
     <Compile Include="Query\OutlineElements1D.cs" />
     <Compile Include="Query\InternalPolyline.cs" />
     <Compile Include="Query\Normal.cs" />
